% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_prob_excl_Omega_raw.R
\name{boot_prob_excl_Omega_raw}
\alias{boot_prob_excl_Omega_raw}
\title{Bootstrap estimation of the probabilities of exclusion and capital Omega}
\usage{
boot_prob_excl_Omega_raw(
  A_int,
  number_Omega_replicates = 1000,
  number_boot_replicates = 1000,
  use_chol_decomp = F
)
}
\arguments{
\item{A_int}{a SxS interaction matrix, where S is the number of species.}

\item{number_Omega_replicates}{a number that specify how many estimations 
of Omega will be calculated by the quasi-Monte Carlo method. By default, 
this parameter is set to 1,000 replicates.}

\item{number_boot_replicates}{a number that specifies how many bootstrap
estimations of the probability of exclusion and Omega will be calculated
for each species. By default, this parameter is set to 1,000 replicates.}

\item{use_chol_decomp}{boolean that specifies if the QR decomposition should
be used to compute Omega.}
}
\value{
A tibble with S rows and 2 x number_boot_replicates columns. Each row 
contains the estimated probabilities of exclusion and values of Omega for 
each species, when considering the intersection of the feasibility domain 
and the unit ball. Specifically, each row contains:
   \itemize{
   \item Resamples of the estimated probabilities of exclusion for species i 
   (prob_excl_rep_x), where x takes values from 1 to number_boot_replicates.
   \item The corresponding value of capital Omega for species i and 
   a given prob_excl_rep_x (Omega_rep_x), where x takes values from 1 to 
   number_boot_replicates.
   }
}
\description{
Given an interaction matrix, this function generates a tibble with S rows,
where S is the number of species in the matrix. Each row contains the
following information for a given: replicates of its probability of exclusion
and the corresponding value of capital Omega for each replicate (i.e., the 
proportion of the feasible parameter #' space inside the unit sphere). Those 
estimations are computed via a quasi-Monte Carlo method and by 
bootstrapping the quasi-Mote Carlo results.
}
\examples{
numCores <- detectCores()
numCores
registerDoParallel(2)
A_int <- -1*diag(c(3,5,7))
boot_prob_excl_Omega_raw(A_int)
stopImplicitCluster()

}
\references{
\url{https://doi.org/10.1016/j.jtbi.2018.04.030}
}
