% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_pairwise_similarity.R
\name{mean_pairwise_similarity}
\alias{mean_pairwise_similarity}
\title{Mean pairwise similarity}
\usage{
mean_pairwise_similarity(probability_vector)
}
\arguments{
\item{probability_vector}{a vector with length S whose elements are the 
average probabilities of extinction of the species in a given community, 
where S is the number of species.}
}
\value{
A non-negative number between 0 and 1 (isotropic community).
}
\description{
Given a vector whose elements are the average probabilities of extinction of
the species in a given community, this function calculates the mean pairwise
similarity of such probabilities.
}
\examples{
probability_vector <- c(.25,.15,.60)
mean_pairwise_similarity(probability_vector)

}
