% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega_bootstrap.R
\name{Omega_bootstrap}
\alias{Omega_bootstrap}
\title{Random estimations of capital Omega}
\usage{
Omega_bootstrap(A_int, replicates = 1000, use_chol_decomp = F)
}
\arguments{
\item{A_int}{a SxS interaction matrix, where S is the number of species.}

\item{replicates}{a number that specifies how many estimations of Omega
will be calculated. By default, this parameter is set to 1,000 replicates.}

\item{use_chol_decomp}{boolean that specifies if the QR decomposition should
be used to compute Omega.}
}
\value{
A vector whose length is equal to the parameter replicate, where each
 element represents an estimation of capital Omega.
}
\description{
Given an interaction matrix, this function generates a vector with 
estimations of capital Omega (i.e., the proportion of the feasible parameter 
space inside the unit sphere). Those estimations are computed via a 
quasi-Monte Carlo method for even relatively large communities.
}
\examples{
numCores <- detectCores()
numCores
registerDoParallel(2)
A_int <- -1*diag(c(3,5,7))
Omega_bootstrap(A_int, replicates = 1e3)
stopImplicitCluster()

}
\references{
\url{https://doi.org/10.1016/j.jtbi.2018.04.030}
}
