% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_omega_iso.R
\name{small_omega_iso}
\alias{small_omega_iso}
\title{Small Omega for the maximum isotropic cap}
\usage{
small_omega_iso(A_int)
}
\arguments{
\item{A_int}{a SxS interaction matrix, where S is the number of species.}
}
\value{
A number that shows the value of value of small Omega, the S root of 
the volume of the maximum isotropic cap, Xi.
}
\description{
Given an SxS interaction matrix, this function computes the value of small 
Omega for maximum isotropic cap (i.e., the S root of the maximum isotropic 
cap, Xi).
}
\examples{
A_int <- -1*diag(c(3,5,7))
small_omega_iso(A_int)

}
